ramebuffer [/]

+++ Engine/Graphics/Shader.hpp
@@ -5,7 +5,6 @@
 
 #include <Asset.hpp>
 #include <Resource.hpp>
-#include <Type.hpp>
 
 namespace Engine {
 
@@ -14,15 +13,15 @@ namespace Engine {
 
     private:
         std::string path;
-        uint16_t shaderType;
+        unsigned shaderType;
 
     public:
         virtual std::shared_ptr<Resource> GetResource() override;
         const std::string &GetPath() const { return path; }
-        uint16_t GetShaderType() const { return shaderType; }
+        unsigned GetShaderType() const { return shaderType; }
 
         void SetPath(const std::string &path) { this->path = path; }
-        void SetShaderType(uint16_t shaderType) { this->shaderType = shaderType; }
+        void SetShaderType(unsigned shaderType) { this->shaderType = shaderType; }
     };
 
     class ENGINE_EXPORT Shader : public Resource {
@@ -38,7 +37,7 @@ namespace Engine {
         virtual void Apply() override;
 
         const std::string &GetPath() const { AShader *ashader = (AShader *)asset; return ashader->GetPath(); }
-        uint16_t GetShaderType() const { AShader *ashader = (AShader *)asset; return ashader->GetShaderType(); }
+        unsigned GetShaderType() const { AShader *ashader = (AShader *)asset; return ashader->GetShaderType(); }
 
         friend class Material;
     };
